03_Scripts/data_generation/requirements.txt
requests
mysql-connector-python
python-dotenvs